+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - Correctly SP out of bounds errors
+- Loading integers can now only be done using the LDI Instruction
 
 ## [0.4.0] - 2020-04-17
 ### Added